requirements.txt
openai
python-dotenv
azure-search-documents
azure-core
gremlinpython
PyPDF2